my_project/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
pydantic==2.4.2
python-dotenv==1.0.0
pytest==7.4.3
mypy==1.6.1
flake8==6.1.0
httpx==0.25.1  # For TestClient
email-validator==2.1.0
